Slz here to fix my own trash code!
I was testing the weather command and noticed that the wind info was being printed as temperature info. For instance:
a wind of 3 kilometers per hour is expected to be printed as: "Kilometer: 3.0 KM/h", but instead it was being printed as: "Celsius: 3.0 Cº"

Turns out I was printing it using the "weather_feels_like_info" template, which causes it to be printed as temperature.

Was I high on drugs when I wrote this code? Like, I tested this command multiple times before and never noticed it.

Literally 2 LOC to solve the problem.